Rearrange standard colors
